+++ repos/views.py
@@ -1,11 +1,13 @@
+from django.contrib import messages
 from django.contrib.auth.mixins import LoginRequiredMixin, PermissionRequiredMixin
 from django.contrib.messages.views import SuccessMessageMixin
 from django.db.models import Q
-from django.urls import reverse_lazy
+from django.db.transaction import on_commit
 from django.views.generic import ListView, CreateView, DetailView, DeleteView
 from django.utils.translation import gettext as _
 
 from repos.models import Repository
+from repos.tasks import clone_remote_repository, remove_local_repository
 import logging
 
 logger = logging.getLogger(__name__)
@@ -41,8 +43,8 @@ class RepositoryList(RepositoryMixin, ListView):
         return Repository.objects.filter(Q(owner=self.owner) | Q(public=True))
 
 
-'''Add a new repo'''
 class RepositoryCreate(RepositoryMixin, CanAdminReposMixin, CreateView):
+    """Add a new repo"""
     success_message = _('Repository was added successfully')
     template_name = 'repository/add.html'
     fields = ['name', 'url', 'username', 'password', 'branches_to_track', 'default_branch', 'public']
@@ -56,7 +58,7 @@ class RepositoryCreate(RepositoryMixin, CanAdminReposMixin, CreateView):
         form.instance.owner = self.request.user
         form.instance.status = Repository.Status.CREATED
         self.object = form.save()
-        # clone_remote_repository.delay(self.object.owner.id, self.project.id, self.object.id)
+        on_commit(lambda: clone_remote_repository.delay(self.object.owner.id, self.object.id))
         return super(RepositoryCreate, self).form_valid(form)
 
 
@@ -67,3 +69,11 @@ class RepositoryDetail(RepositoryMixin, CanSeeRepoMixin, DetailView):
 class RepositoryDelete(RepositoryMixin, CanAdminReposMixin, DeleteView):
     success_message = _('Repository was deleted successfully')
     success_url = '/repos/'
+
+    def delete(self, request, *args, **kwargs):
+        self.owner = request.user
+        result = super(RepositoryDelete, self).delete(request, *args, **kwargs)
+
+        remove_local_repository.delay(self.object.owner.id, self.object.name)
+        messages.success(self.request, self.success_message)
+        return result